github_prometheus_exporter/pyproject.toml
[tool.poetry]
name = "github-prometheus-exporter"
version = "0.1.0"
description = ""
authors = ["hxnir <nir@hakasecurity.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pygithub = "^2.1.1"
ipython = "^8.20.0"
ruff = "^0.1.13"
mypy = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120
target-version = "py311"
preview = true
select = [
    "ALL"
]
ignore = [
    "A003", # Class attribute is shadowing a Python builtin
    "ANN101", # Missing type annotation for `self` in method
    "ANN102", #  Missing type annotation for `cls` in classmethod
    "ANN401", # Dynamically typed expressions (typing.Any) are disallowed
    "COM812", # missing-trailing-comma
    "D", # Docstring
    "DTZ", # The use of `datetime.datetime.now()` without `tz` argument is not allowed. TODO: Remove this.
    "EXE002", # The file is executable but no shebang is present
    "FBT001", # Boolean-typed positional argument in function definition
    "FBT002", # Boolean default positional argument in function definition
    "FIX002", # Line contains TODO, consider resolving the issue
    "G004", # Logging statement uses f-string
    "PGH003", # Use specific rule codes when ignoring type issues (Dont use type: ignore)
    "PTH", # `open()` should be replaced by `Path.open()`. TODO: Remove this.
    "SLOT000", # Subclasses of `str` should define `__slots__`
    "TD", # Missing author in TODO
    "W605", # adds `r` before string where backslash exists. TODO: remove this.
    "FBT003", # Boolean positional value in function call. TODO: remove this.
    "N818",  # Exception name should be named with an Error suffix. TODO: remove this.
    "CPY001", #  Missing copyright notice at top of file
    "PLR6201", # Use a `set` literal when testing for membership
    "PLW1514", # `open` in text mode without explicit `encoding` argument TODO: remove this.
    "PLC1901", # s != "" can be simplified to s as an empty string is falsey. Has many false positives.
    "PLR0917", # Too many positional arguments. TODO: remove this.
    "PLR6301", # Method could be a function, class method, or static method.
    "ISC001" # single-line-implicit-string-concatenation, conflicts with ruff formating.
]

[tool.ruff.pylint]
max-args = 7
max-branches = 7

[tool.ruff.mccabe]
max-complexity = 8

[tool.ruff.isort]
known-first-party = ["haka_schema", "haka_libraries", "haka_dlp_analyzer_api_client"]